App/requirements.txt
pip
python-telegram-bot
